backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
langchain>=0.1.0
langchain-core
langchain-groq>=0.1.0
pypdf>=4.0.0
python-docx>=1.1.0
pydantic>=2.8.0
langchain-community
python-dotenv
